fix: Handle bluetooth status in a more robust way
The current implementation will always return 0 (=disabled) as bluetooth
status, if the device is already registered. This is happens because the
uart is not present as device in /dev (bound by driver).

This refactor the index calculation for rfkill and check if the bluetooth
device is already present. The device is still being tried to initialize,
if the hci device is not present.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>